Move `WeakRef` to `CoreBind`, since it shouldn't be (and isn't) used internally.
